requirements.txt
requests
pandas
matplotlib
pyrogram
tgcrypto
python-dotenv
